refactor config mod
