+++ src-tauri/src/bin/tester.rs
@@ -19,7 +19,7 @@ fn main() {
     trace!("App started.");
     info!("Logging level: TRACE");
 
-    rulaub_backend::config::loader::load_and_watch_config_file();
+    rulaub_backend::config::utils::load_and_watch_config_file();
 
     tracing_handle.reload(create_env_filter("TRACE")).unwrap();
