+++ src-tauri/src/config/file.rs
@@ -1,26 +1,22 @@
 use std::{
     fs::{create_dir_all, File},
-    io::{Error as IOError, Write},
+    io::Write,
     path::Path,
 };
 
-use thiserror::Error;
-use tracing::{debug, error, trace};
+use std::{sync::mpsc::channel, time::Duration};
 
-use super::DEFAULT_CONFIG_TOML_NICE_STR;
+use configlib::File as ConfigFile;
+use notify::{DebouncedEvent, RecommendedWatcher, RecursiveMode, Watcher};
+
+use super::{
+    create_default_config,
+    utils::{log_config, ConfigFileError, FileWatchError},
+    CONFIG, CONFIG_FILE_PATH, DEFAULT_CONFIG_TOML_NICE_STR,
+};
 use crate::{NL, PROJECT_DIRS};
 
-#[derive(Error, Debug)]
-pub enum ConfigFileError {
-    #[error("Could not find project dirs.")]
-    ProjectDirsNotFound,
-    #[error("Could not stringify config file path.")]
-    PathStringify,
-    #[error("Could not interact with config file.")]
-    FileError(#[from] IOError),
-}
-
-fn get_conf_file_path() -> Result<String, ConfigFileError> {
+pub fn get_conf_file_path() -> Result<String, ConfigFileError> {
     match &*PROJECT_DIRS {
         None => {
             error!("Could not find project directories for this OS.");
@@ -51,10 +47,6 @@ pub fn check_path_is_file(path: &str) -> bool {
     is_file
 }
 
-lazy_static! {
-    pub static ref CONFIG_FILE_PATH: String = get_conf_file_path().unwrap_or_default();
-}
-
 #[tracing::instrument]
 pub fn write_to_config_file(content: &str) -> Result<(), ConfigFileError> {
     trace!("Write to config file.{}Path: {:?}", NL, *CONFIG_FILE_PATH);
@@ -74,4 +66,56 @@ pub fn create_config_file_with_defaults() -> Result<(), ConfigFileError> {
     trace!("Create new config file with defaults.");
     write_to_config_file(&DEFAULT_CONFIG_TOML_NICE_STR)?;
     Ok(())
+}
+
+#[tracing::instrument]
+pub fn load_config_file() {
+    debug!("Load config file.");
+    let mut config_guard = CONFIG.write();
+    if let Err(err) = config_guard.merge(ConfigFile::with_name(&CONFIG_FILE_PATH)) {
+        error!(
+            message = concat!(
+                "Failed to merge config file into config struct. ",
+                "Settings config struct back to default config."
+            ),
+            error = ?err
+        );
+        // TODO:#i# send msg to err frontend saying to fix config and restart
+        *config_guard = create_default_config();
+    };
+    log_config();
+}
+
+#[tracing::instrument]
+pub fn watch_config_file() -> Result<(), FileWatchError> {
+    trace!("Start watching config file.");
+    let (tx, rx) = channel();
+
+    let mut watcher: RecommendedWatcher = Watcher::new(tx, Duration::from_secs(2))?;
+    watcher.watch(&CONFIG_FILE_PATH[..], RecursiveMode::NonRecursive)?;
+
+    loop {
+        match rx.recv() {
+            Ok(DebouncedEvent::Write(_)) => {
+                info!("config.toml updated. Refreshing configuration.");
+                let mut config_guard = CONFIG.write();
+                if let Err(err) = config_guard.refresh() {
+                    error!(
+                        message = concat!(
+                            "Failed to update config. ",
+                            "Probably invalid config file content."
+                        ),
+                        error = ?err
+                    );
+                    return Err(FileWatchError::ConfigUpdateError);
+                }
+                log_config();
+            }
+            Err(err) => error!(
+                message = "Error while watching config file.",
+                err = ?err
+            ),
+            _ => { /* Ignore other event */ }
+        }
+    }
 }
